Moved logic to model-controller
