+++ EiDiA_backend/src/routes/upload.js
@@ -5,8 +5,9 @@ const router = express.Router();
 
 const middleWares = require('../middlewares');
 const UploadController = require('../controllers/upload');
+const ModelController = require('../controllers/model');
 
-router.get('/attributes/get/:selectedDocumentTypeId', middleWares.checkAuthentication, UploadController.listAttributes);
+router.get('/attributes/get/:selectedDocumentTypeId', middleWares.checkAuthentication, ModelController.listAttributes);
 
 router.post('/document', middleWares.checkAuthentication, UploadController.uploadDocument);
 router.post('/data', middleWares.checkAuthentication, UploadController.addAttributes);
